companySevice 2개로 분리
